Add files via upload
Supplement the hand gesture similarity annotation data. The inter-class similarity is measured by "similar_skeleton", while the intra-class similarity is measured by "Standard_skeleton" as the standard data. The corresponding 10 samples in "ISW-1000" are used as the comparison data. The manually annotated intra-class similarity scores are in the file "one_ten_similar.xlsx".